refactor: 문서함 vault 도구를 설정 메뉴로 합병 (B2, 중복 표면 제거)
문서함 헤더의 VaultToolsMenu 드롭다운을 AppSettingsMenu 로 흡수해 같은
관심사(AI agent 설정·검증)의 집을 하나로 통일한다. LNB 는 늘리지 않았고
신규 탭·신규 표면 0.

이관(설정 메뉴로):
- AI agent 설정 패널 전체 → 새 위젯 VaultAgentSetupPanel, 설정 → MCP/Agents 탭.
  설정 파일 상태(.mcp.json/.codex/.mcp.json.example 존재·유효), 수리
  (agent-setup --write), 복사 패킷 8종(설정 패킷·검증 프롬프트·CLI runbook·
  첫 연결 증거·JSON gate·post-change sync·MCP JSON·Codex TOML·codex mcp add·
  설정 상태/repair CLI), mode chooser, 검증 게이트, 3단계 체크리스트, proof
  rows, first-contact 계약 rows, mcp-verify/json-gate 미리보기.
- LocalVaultPicker(폴더 열기·닫기·새로고침·Finder·최근·권한 복구) → 설정 →
  Workspace 탭.

제거(중복):
- mcpAgents 탭: vault 로드 시 vault-aware 패널만 노출하고 정적 first-calls
  교육 블록은 숨김(둘 다 first-contact 패킷을 다뤄 중복).
- mcpAgents 의 정적 state-decision-table 삭제 — verification 탭의 동일 state
  ladder 와 중복(같은 mcpStateRows 배열).

유지(문서함 헤더):
- vault pill "vault 바꾸기" = 고빈도 swap 만 유지 → 네이티브 폴더 재선택
  (openLocalVault). 데스크톱 샘플→로컬 전환은 source 전환 직결.
- vault 도구 게어 → 설정 게어(AppSettingsMenu)로 교체.
- vault chip 팝오버에 "도구는 설정으로 이동" 브리지 한 줄(ko/en, 이번 릴리스 한정).
- 에러/권한 배너: requestPermission()/폴더 재선택으로 직접 결선.

정리:
- VaultToolsMenu.tsx + 테스트 삭제. 테스트 26건은 VaultAgentSetupPanel.test.tsx
  로 이식(picker 전용 2건은 LocalVaultPicker.test 가 이미 커버해 제외). 설정
  메뉴 테스트에 picker 도달 + 패널 non-render 케이스 추가.
- DocsVaultPage 결선 정리(dead advancedOpen/ref/localVaultValidationSummary/
  Settings2 제거), docs-vault 배럴에서 VaultToolsMenu export 제거.
- 컨텍스트: AppSettingsMenu 가 useLocalVault() 로 app-wide LocalVaultProvider
  를 직접 읽어 어느 페이지 헤더에서도 같은 vault 를 본다(props 주입 불필요).
- dogfood 온톨로지 evidence(agent-config-onboarding/agent-graph-readiness)
  갱신 + 매니페스트 재생성, FEATURES/DEVELOPMENT-CHECKS/CHANGELOG 현행화.

검증: vitest(app-settings-menu·docs-vault 176 + 전체 2538 pass) · tsc · eslint ·
i18n messages · vault:validate · vault:audit(drift 0) · build(static export).
