+++ messages/en.json
@@ -71,6 +71,8 @@
       "vaultBodyStatic": "Use the macOS app or local workspace flow before selecting a writable folder.",
       "vaultCtaLocal": "Open workspace",
       "vaultCtaStatic": "Start local workspace",
+      "vaultRevealError": "Could not open Finder: {message}",
+      "vaultRevealErrorFallback": "Could not open the vault in Finder.",
       "agentTitle": "AI agent connection",
       "agentBody": "Review MCP settings, tools/list proof, and graph DB gates on the connection and verification screen.",
       "agentCta": "Open verification",
@@ -1785,6 +1787,7 @@
       "sourceAriaLabel": "Workspace source",
       "vaultToolsTooltip": "Workspace tools",
       "vaultToolsAriaLabel": "Open workspace tools menu",
+      "vaultToolsMovedHint": "AI agent setup and verification tools moved to Settings — open the gear (⚙) at the top right.",
       "contractToggleLabel": "Workspace checks",
       "contractToggleHide": "Close workspace checks",
       "contractToggleShow": "Show workspace checks",
