+++ messages/ko.json
@@ -71,6 +71,8 @@
       "vaultBodyStatic": "저장 가능한 폴더를 선택하려면 macOS 앱 또는 로컬 작업공간 흐름으로 이동합니다.",
       "vaultCtaLocal": "작업공간 열기",
       "vaultCtaStatic": "로컬 작업공간 시작",
+      "vaultRevealError": "Finder 열기 실패: {message}",
+      "vaultRevealErrorFallback": "Finder에서 vault를 열지 못했습니다.",
       "agentTitle": "AI 에이전트 연결",
       "agentBody": "MCP 설정, tools/list 증거, 그래프 검증 게이트는 연결·검증 화면에서 확인합니다.",
       "agentCta": "연결·검증 열기",
@@ -1785,6 +1787,7 @@
       "sourceAriaLabel": "문서함 선택",
       "vaultToolsTooltip": "문서함 도구",
       "vaultToolsAriaLabel": "문서함 도구 메뉴 열기",
+      "vaultToolsMovedHint": "AI agent 설정·검증 도구는 설정으로 이동했어요 — 오른쪽 위 설정(⚙)에서 엽니다.",
       "contractToggleLabel": "문서함 점검",
       "contractToggleHide": "문서함 점검 닫기",
       "contractToggleShow": "문서함 점검 보기",
