+++ src/widgets/docs-vault/index.ts
@@ -1,5 +1,4 @@
 export { DocsVaultTree } from './ui/DocsVaultTree';
-export { VaultToolsMenu } from './ui/VaultToolsMenu';
 export { DocsVaultViewer } from './ui/DocsVaultViewer';
 export { DocsVaultEditor } from './ui/DocsVaultEditor';
 export { DocsVaultUnifiedPalette } from './ui/DocsVaultUnifiedPalette';
